feat(workers): implement worker architecture with BullMQ and Redis\n\nIntroduced a new worker system using BullMQ for job processing. Added Dockerfile for worker deployment, created necessary TypeScript configuration, and established a shared service for job management. Removed outdated queue implementation and refactored job addition logic to enhance clarity and maintainability.
