+++ .gitignore
@@ -22,6 +22,7 @@ backups
 
 # production
 /build
+/apps/workers/dist
 
 # misc
 .DS_Store
