--- lib/queue.ts
@@ -1,49 +1,0 @@
-import { JobsOptions, Queue } from "bullmq"
-import IORedis from "ioredis"
-
-/*
-  A thin wrapper around BullMQ that provides a singleton Queue instance which
-  can be shared across the Next.js application _and_ any external worker
-  process. The queue persists all jobs in Redis, allowing workers to live
-  independently of the Next.js web server lifecycle.
-*/
-
-// TODO: Move this file to a shared library / service when we have the architecture set up for it.
-// Since this will likely be used by multiple services.
-
-// Lazily-initialised Redis connection – re-using the same connection prevents
-// BullMQ from spawning extra listeners and helps keep the number of open file
-// descriptors low in serverless environments.
-let connection: IORedis | null = null
-function getRedisConnection(): IORedis {
-  if (connection) return connection
-
-  const redisUrl =
-    process.env.REDIS_URL ||
-    process.env.REDIS_CONNECTION_STRING ||
-    "redis://localhost:6379"
-
-  connection = new IORedis(redisUrl)
-  return connection
-}
-
-const queues = new Map<string, Queue>()
-
-export function getQueue(name: string): Queue {
-  if (queues.has(name)) return queues.get(name) as Queue
-
-  const queue = new Queue(name, { connection: getRedisConnection() })
-  queues.set(name, queue)
-
-  return queue
-}
-
-export async function addJob<T extends Record<string, unknown>>(
-  name: string,
-  data: T,
-  opts: JobsOptions = {}
-): Promise<string> {
-  const queue = getQueue(name)
-  const job = await queue.add(name, data, opts)
-  return job.id as string
-}
