+++ tsconfig.json
@@ -4,7 +4,8 @@
     "paths": {
       "@/styles/*": ["styles/*"],
       "@/components/*": ["components/*"],
-      "@/*": ["./*"]
+      "@/*": ["./*"],
+      "@shared/*": ["shared/src/*"]
     },
     "target": "ES2017",
     "lib": ["dom", "dom.iterable", "esnext"],
